Another reorganization
This reorganization makes it clearer where code is coming from and
avoids a lot of the issues from the earlier reorganizations. In
particular, there were several cases where Python3 was confused
about the top-level modules in use. By breaking out the invocation
point from the modules, this makes it far clearer.